module_paper_search/submodule_metadata_embedding/requirements.txt
torch>=1.10.0
transformers>=4.15.0
faiss-cpu>=1.7.2
numpy>=1.20.0
tqdm>=4.62.0
